Move stuff that depends on rocket + testchipip out
Other misc. cleanup
